# openapi.yaml
components:
  schemas:
...
    CreateCardReq:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/CreateCardReq.json
          format: uri
          readOnly: true
          type: string
        back:
          type: string
        front:
          type: string
      type: object
...
    CreateDeckReq:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/CreateDeckReq.json
          format: uri
          readOnly: true
          type: string
        color:
          $ref: "#/components/schemas/UserColor"
        name:
          type: string
        workspaceId:
          type: string
      type: object
...
    CreateQuizReq:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/CreateQuizReq.json
          format: uri
          readOnly: true
          type: string
        chapters:
          items:
            type: string
          nullable: true
          type: array
        name:
          type: string
        privacy:
          $ref: "#/components/schemas/Privacy"
        questions:
          items:
            additionalProperties: {}
            type: object
          nullable: true
          type: array
        timeLimitMin:
          format: int64
          type: integer
        workspaceId:
          type: string
      type: object
...
    Deck:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/Deck.json
          format: uri
          readOnly: true
          type: string
        cardCount:
          format: int64
          type: integer
        color:
          $ref: "#/components/schemas/UserColor"
        dueCount:
          format: int64
          type: integer
        id:
          type: string
        knownPct:
          format: int64
          type: integer
        name:
          type: string
        workspaceId:
          type: string
        workspaceName:
          type: string
      required:
        - id
        - name
        - workspaceId
        - workspaceName
        - color
        - cardCount
        - knownPct
        - dueCount
      type: object
...
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          example: Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          nullable: true
          type: array
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          example: https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          example: 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          example: Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          example: https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    Flashcard:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/Flashcard.json
          format: uri
          readOnly: true
          type: string
        back:
          type: string
        deckId:
          type: string
        front:
          type: string
        id:
          type: string
        known:
          type: boolean
        srs:
          $ref: "#/components/schemas/SrsState"
      required:
        - id
        - deckId
        - front
        - back
        - known
        - srs
      type: object
...
    Quiz:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/Quiz.json
          format: uri
          readOnly: true
          type: string
        chapters:
          items:
            type: string
          type: array
        createdAt:
          format: date-time
          type: string
        id:
          type: string
        name:
          type: string
        privacy:
          $ref: "#/components/schemas/Privacy"
        questions:
          items:
            additionalProperties: {}
            type: object
          type: array
        timeLimitMin:
          format: int64
          type: integer
        workspaceId:
          type: string
        workspaceName:
          type: string
      required:
        - id
        - name
        - workspaceId
        - workspaceName
        - chapters
        - questions
        - createdAt
        - privacy
      type: object
...
    UserColor:
      enum:
        - green
        - purple
        - blue
        - amber
        - coral
        - graphite
        - transparent
      type: string
    Workspace:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/Workspace.json
          format: uri
          readOnly: true
          type: string
        chapterCount:
          format: int64
          type: integer
        color:
          $ref: "#/components/schemas/UserColor"
        createdAt:
          format: date-time
          type: string
        fileCount:
          format: int64
          type: integer
        id:
          type: string
        lastAccessedAt:
          format: date-time
          type: string
        name:
          type: string
        privacy:
          $ref: "#/components/schemas/Privacy"
        tags:
          items:
            $ref: "#/components/schemas/Tag"
          type: array
      required:
        - id
        - name
        - color
        - privacy
        - tags
        - chapterCount
        - fileCount
        - createdAt
        - lastAccessedAt
      type: object
...
info:
  title: Evo Notes API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/decks:
    get:
      operationId: listDecks
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Deck"
                nullable: true
                type: array
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List decks
      tags:
        - Flashcards
    post:
      operationId: createDeck
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeckReq"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Create a deck
      tags:
        - Flashcards
  /api/decks/{id}:
    get:
      operationId: getDeck
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get a deck
      tags:
        - Flashcards
  /api/decks/{id}/cards:
    get:
      operationId: listCards
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Flashcard"
                nullable: true
                type: array
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List cards in a deck
      tags:
        - Flashcards
    post:
      operationId: createCard
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardReq"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Create a card
      tags:
        - Flashcards
...
  /api/quizzes:
    get:
      operationId: listQuizzes
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Quiz"
                nullable: true
                type: array
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List quizzes
      tags:
        - Quizzes
    post:
      operationId: createQuiz
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizReq"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Create a quiz
      tags:
        - Quizzes
...
  /api/workspaces/{id}:
    delete:
      operationId: deleteWorkspace
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Delete a workspace
      tags:
        - Workspaces
    get:
      operationId: getWorkspace
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get a workspace
      tags:
        - Workspaces
    patch:
      operationId: updateWorkspace
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceReq"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Update a workspace
      tags:
        - Workspaces
  /api/workspaces/{id}/chapters:
    get:
      operationId: listChapters
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Chapter"
                nullable: true
                type: array
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List chapters
      tags:
        - Content
    post:
      operationId: addChapter
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddChapterReq"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Add a chapter
      tags:
        - Content
...
  /api/workspaces/{id}/generate:
    post:
      operationId: generateMaterial
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOptions"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResult"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Generate study material from workspace scope
      tags:
        - Materials
...